Stash work
While this approach seems promising, I ultimately stumbled
on serious problems with FBOs diminishing alpha values for
no discernible reason. Then there also was the unsolved
problem of constructive and destructive interference and
how to properly blend things together. Ultimately, it was
suggested to me to go back to a voronoi-based solution so
I'll probably do just that now. I have a few ideas on how
that could be implemented in practice.