Offer a flat color shading mode
This has a more modern futuristic look to it - I quite like it I'll be honest.
It needs some edge detection to look good as otherwise the empire borders
don't completely align well, but otherwise it's on a good trajectory.

Another thing that might need to be adjusted is the distance weighting formula,
as it stands it at times tends to completely ignore smaller blobs of territory,
so we might need to more agressively increase the depth proportional to distance
to the center.