Reduce amount of mixins targetting galimulator classes
Prefer the usage of SLAPI methods instead. The remaining mixin is
a bit of a special child, but might be replaced by something else
in the long term. Probably an event?

As a logical conclusion to this change, the required version of
SLAPI shot up drastically - almost by a year. Goes to show that
I haven't been too active recently. Oh well - life moves on.